requirements.txt
anyascii==0.3.1
asgiref
attrs==23.1.0
autopep8==1.6.0
beautifulsoup4==4.11.1
certifi==2021.10.8
cffi==1.16.0
charset-normalizer==2.0.9
click==8.1.6
colorama==0.4.6
crispy-bootstrap5==0.6
cryptography==36.0.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.7
Django==3.2.9
django-admin-volt==1.0.10
django-allauth==0.61.1
django-crispy-forms==1.13.0
django-decouple==2.1
django-extensions==3.1.3
django-filter==22.1
django-import-export==3.2.0
django-js-asset==2.0.0
django-leaflet==0.29.0
django-modelcluster==6.0
django-mptt==0.13.4
django-permissionedforms==0.1
django-rest-framework==0.1.0
django-taggit==3.0.0
django-treebeard==4.5.1
djangorestframework==3.13.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
exceptiongroup==1.1.2
ez-setup==0.9
filelock==3.12.4
fontawesomefree==6.2.0
future==0.18.3
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
html5lib==1.1
idna==3.3
l18n==2021.3
lxml==4.9.3
MarkupPy==1.14
numpy==1.26.0
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.10
outcome==1.2.0
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
pipenv==2023.7.23
platformdirs==3.11.0
psycopg2==2.9.2
psycopg2-binary==2.9.9
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.5
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
ratelim==0.1.6
requests==2.26.0
requests-oauthlib==1.3.0
selenium==4.13.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.4.0
telepath==0.3
toml==0.10.2
tqdm==4.65.0
trio==0.22.1
trio-websocket==0.10.3
tzdata==2023.3
urllib3==1.26.7
virtualenv==20.24.5
virtualenv-clone==0.5.7
wagtail==4.0.4
webdriver-manager==3.8.6
webencodings==0.5.1
Willow==1.4.1
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
